ClienteController REST creado. Redirecciones y paths modificados. Para mas data de rest revisar https://spring.io/guides/gs/rest-service/
