+++ src/main/java/com/fra/clientes/controller/rest/ClientesController.java
@@ -1,74 +1,37 @@
-package com.fra.clientes.controller;
-
-
+package com.fra.clientes.controller.rest;
 
 import java.util.List;
 
 import javax.servlet.http.HttpServletRequest;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Controller;
+import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
-import org.springframework.web.servlet.ModelAndView;
+import org.springframework.web.bind.annotation.ResponseStatus;
+import org.springframework.web.bind.annotation.RestController;
 
 import com.fra.clientes.models.Cliente;
 import com.fra.clientes.services.ClienteService;
 
-@Controller
+@RestController
 public class ClientesController {
-
+	
 	@Autowired
 	private ClienteService clienteService;
-
-	/**
-	 * SPRING MVC
-	 * 
-	 */
-	@RequestMapping(value = "/home" , method = RequestMethod.GET)
-	public ModelAndView home() {
-		
-		//TODO: MANEJO LA RESPUESTA DE JSON
-		List<Cliente> clientes = cliente();
-		
-		ModelAndView modelAndView = new ModelAndView("Home");
-		modelAndView.addObject("Lista", clientes);
-		
-		return modelAndView;
-	}
-
-	@RequestMapping(value = "/cliente/form/add" , method = RequestMethod.GET)
-	public ModelAndView formAddCliente() {
-		ModelAndView modelAndView = new ModelAndView("NewCliente");
-		return modelAndView;
-	}
-	
-	@RequestMapping(value = "/cliente/{id}/form/update" , method = RequestMethod.GET)
-	public ModelAndView formUpdateCliente(@PathVariable("id") long id) {
-		
-		Cliente cliente = clienteService.getClienteById(id);
-		ModelAndView modelAndView = new ModelAndView("EditCliente");
-		modelAndView.addObject("Cliente", cliente);
-		
-		return modelAndView;
-	}	
-
-
-	/**
-	 * REST API
-	 * 
-	 */
 	
 	@RequestMapping(value = "/cliente" , method = RequestMethod.GET)
+	@ResponseStatus(HttpStatus.OK)
 	public List<Cliente> cliente() {
-		//TODO: DEVOLVER UN JSON
+		//TODO: DEVOLVER UN JSON, BUSCAR JACKSON
 		return clienteService.getClientes();
 	}
 	
 
 	@RequestMapping(value = "/cliente", method = RequestMethod.POST)
-	public String saveOrUpdate(HttpServletRequest request) {
+	@ResponseStatus(HttpStatus.CREATED)
+	public void addCliente(HttpServletRequest request) {
 
 		//TODO: RECIBIR UN JSON
 		
@@ -80,14 +43,13 @@ public class ClientesController {
 		Cliente cliente = new Cliente(nombreApellido,telefono,direccion,establecimiento);
 		
 		clienteService.addCliente(cliente);
-		
-		return "redirect:home";
 	}
 		
-	@RequestMapping(value = "/Editar Cliente", method = RequestMethod.POST)
-	public String editCliente(HttpServletRequest request) {
+	@RequestMapping(value = "/cliente/{id}", method = RequestMethod.PATCH)
+	@ResponseStatus(HttpStatus.NO_CONTENT)
+	public void editCliente(HttpServletRequest request) {
 
-		//TODO: USAR EL SAVE OR UPDATE, BORRAR ESTE METODO
+		//TODO: RECIBIR UN JSON
 		
 		long id = Long.valueOf(request.getParameter("id"));
 		String nombreApellido = request.getParameter("nombreApellido");
@@ -100,17 +62,13 @@ public class ClientesController {
 		cliente.setId(id);
 		
 		clienteService.updateCliente(cliente);
-				
-		return "redirect:home";
 	}	
 	
 
-	@RequestMapping(value = "/cliente/{id}/delete", method = RequestMethod.POST)
-	public String deleteCliente(@PathVariable("id") long id) {
-		
-		//TODO: SOLO RETORNAR STATUS CODE
+	@RequestMapping(value = "/cliente/{id}", method = RequestMethod.DELETE)
+	@ResponseStatus(HttpStatus.NO_CONTENT)
+	public void deleteCliente(@PathVariable("id") long id) {		
 		this.clienteService.deleteClienteById(id);
-		return "redirect:/home";
 	}
 
 }
